adding api call to get sales orders in a given range
udating the api to be able to return salesorders between a given range,
as well as the client/shell to make use of this.

Has resulted in a lot of duplicate code, but next patch will go through
and clean up a bunch of that, and extract a few validators into either decorators,
or at least helper functions.

Change-Id: I59ed58c32b8dd4a75babf70df188e965e5e94a16