Add new transformer for network service
The network service pollsters of Ceilometer are using
status as the volume(see https://github.com/openstack/ceilometer/
blob/master/ceilometer/network/services/vpnaas.py#L55), so we have
to check the volume to make sure only the active service is
charged(0=inactive, 1=active). The patch is adding a new transformer
based on above statement and a test cause for the code change.

Story: #14112

Change-Id: I9d2e435e3d58b7d864f6b27d4dc2acdece648d61