fpm: Convert deb-pre-depends to normal depends.
This makes the package install a little less weird, esp with a loose
deb. Previously, dpkg would refuse to install the package at all, if
any dependency was unmet. Now, you can install in a broken state, and
then `apt-get install -f` or similar to fix up the missing dependencies.

Change-Id: I0089934db6c009c5e158260785d5af677719a38f