Fix tagging issue after instance deleted
Now we're tagging the windows instance even after the instance
has already existed in Distil. This may cuase problem if the
instance is deleted in the middle of metrics collecting period.
For example, if a Windows instance A is created at 1:30PM and
Distil collects the usage at 2:00PM, then user delete the
instance at 2:30PM, then when Distil collect usage at 3:00PM,
bad things happened. The instance has been deleted, but there
are still samples between 2:00PM - 2:30PM, and when Distil try
to check if the instance is a Windows, it can't get any info.
As a result, the instance will be tagged as unknown. This patch
fixes it.

Change-Id: I46f61d1a4a858d1ecfcd0820e541030c927765a5