Remove odoo glue script
We're adding more internal logic to this script and obviously it
is not suitable to keep it in this repo anymore. In the future,
it would be nice if we can make the quote as an API endpoint and
put the odoo glue script logic as a part of that when using odoo
as backend.

Change-Id: I09a2ec0459f14903f6ac5f2a01a25a2f494e0b84