Fix long resource id issue
Currently the column size of resource id in DB is 100 chars, but
the container name of swift could be 256, plus project id and a
'/', the id for a swift container could be 32+1+256. So this is
a fix for the problem. But instead of checking the length of
resource id, here I'm hashing the name only for swift to get
a consistent id for swift billing. Another change will be
proposed to openstack-billing to handle this case as well.

Change-Id: I9ced37d21f67305c065e61ecf25621d583db233e