Fix last_collect time for new project
When a new project is created in Keystone, the distil-collector service
will collect its usage from the pre-configured 'dawn_of_time' which is
a very early time.

This patch will get the earliest last collect time from all existing
projects and use that as the start time for the new project.

Change-Id: I02d3c4cb1c36ca043aaac0c2c25965eb1f5afa15