Improve cache about get products
For most of the cases of Distil API, the request comes
from billing panel. Billing panel sends 1 API call for /invoices and
several API calls for /quotations against different regions. So it's
not efficient to specify the region for get_products method because
it won't help cache the products based on the parameters.

Change-Id: I2f7e2ce8c3f43e79b9e9084903fca1bc4f44f82a